+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # More Plates Changelog
 
+### 1.12.2-5.0.2
+- Updated config (I recommend delete old config!)
+- Added Ender IO Endergy support
+
 ### 1.12.2-5.0.1
 - Updated to new Actually Additions API
 - Changed glowstone_gear to refined_glowstone_gear and glowstone_plate to refined_glowstone_plate